pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polyllm"
version = "1.0.5"
authors = [
    { name="Rory Butler" },
]
description = "A unified interface for querying multiple large language models"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "backoff",
    "pydantic",
]

[project.optional-dependencies]
all = [
    "llama-cpp-python",
    "mlx-lm",
    "ollama",
    "openai",
    "google-generativeai",
    "anthropic",
    "litellm",
]

[project.urls]
"Homepage" = "https://github.com/RoryMB/polyllm"
"Bug Tracker" = "https://github.com/RoryMB/polyllm/issues"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "opencv-contrib-python",
    "pillow",
    "pytest",
]
